Start conversion
