+++ torch/nn/functional.py
@@ -14,6 +14,7 @@ from ._functions import vision
 from ._functions.thnn.fold import Col2Im, Im2Col
 from .modules.utils import _single, _pair, _triple, _list_with_default
 from . import grad
+from .._jit import weak_script
 
 _VF = torch._C._VariableFunctions
 
@@ -906,6 +907,7 @@ def hardshrink(input, lambd=0.5):
     return torch.hardshrink(input, lambd)
 
 
+@torch._jit.weak_script
 def tanhshrink(input):
     r"""tanhshrink(input) -> Tensor
 
@@ -916,6 +918,7 @@ def tanhshrink(input):
     return input - input.tanh()
 
 
+@torch._jit.weak_script
 def softsign(input):
     r"""softsign(input) -> Tensor
